refactor related-information
