+++ website/src/components/related-information/index.js
@@ -1,0 +1,1 @@
+export { RelatedInformation } from './related-information';
